refactor: renamed ContentInEvidence block in HighlitedContent block
