+++ locales/nl/LC_MESSAGES/volto.po
@@ -116,6 +116,7 @@ msgid "Content"
 msgstr ""
 
 #: components/ItaliaTheme/Blocks/ContentInEvidence/Sidebar
+#: components/ItaliaTheme/Blocks/HighlightedContent/Sidebar
 msgid "Content in evidence"
 msgstr ""
 
@@ -207,11 +208,13 @@ msgid "Link to"
 msgstr ""
 
 #: components/ItaliaTheme/Blocks/ContentInEvidence/Sidebar
+#: components/ItaliaTheme/Blocks/HighlightedContent/Sidebar
 #: components/ItaliaTheme/Blocks/NewsHome/Sidebar
 msgid "LinkMore"
 msgstr ""
 
 #: components/ItaliaTheme/Blocks/ContentInEvidence/Sidebar
+#: components/ItaliaTheme/Blocks/HighlightedContent/Sidebar
 #: components/ItaliaTheme/Blocks/NewsHome/Sidebar
 msgid "Linkto title"
 msgstr ""
@@ -654,6 +657,7 @@ msgid "bando_vedi"
 msgstr ""
 
 #: components/ItaliaTheme/Blocks/ContentInEvidence/Sidebar
+#: components/ItaliaTheme/Blocks/HighlightedContent/Sidebar
 msgid "bg_color"
 msgstr ""
 
@@ -725,10 +729,12 @@ msgid "closeSearch"
 msgstr ""
 
 #: components/ItaliaTheme/Blocks/ContentInEvidence/Sidebar
+#: components/ItaliaTheme/Blocks/HighlightedContent/Sidebar
 msgid "color_primary"
 msgstr ""
 
 #: components/ItaliaTheme/Blocks/ContentInEvidence/Sidebar
+#: components/ItaliaTheme/Blocks/HighlightedContent/Sidebar
 msgid "color_transparent"
 msgstr ""
 
@@ -972,6 +978,7 @@ msgid "email_sede"
 msgstr ""
 
 #: components/ItaliaTheme/Blocks/ContentInEvidence/Edit
+#: components/ItaliaTheme/Blocks/HighlightedContent/Edit
 #: components/ItaliaTheme/Blocks/NewsHome/Edit
 msgid "emptySelection"
 msgstr ""
@@ -1253,6 +1260,7 @@ msgid "modulo_formati_alternativi"
 msgstr ""
 
 #: components/ItaliaTheme/Blocks/ContentInEvidence/Sidebar
+#: components/ItaliaTheme/Blocks/HighlightedContent/Sidebar
 #: components/ItaliaTheme/Blocks/NewsHome/Sidebar
 #: config/Blocks/ListingOptions/cardWithImageTemplate
 msgid "natural_image_size"
@@ -1881,6 +1889,7 @@ msgid "show_only_first_ribbon"
 msgstr ""
 
 #: components/ItaliaTheme/Blocks/ContentInEvidence/Sidebar
+#: components/ItaliaTheme/Blocks/HighlightedContent/Sidebar
 #: config/Blocks/ListingOptions/utils
 msgid "show_section"
 msgstr ""
@@ -1895,6 +1904,7 @@ msgid "show_topics"
 msgstr ""
 
 #: components/ItaliaTheme/Blocks/ContentInEvidence/Sidebar
+#: components/ItaliaTheme/Blocks/HighlightedContent/Sidebar
 #: config/Blocks/ListingOptions/utils
 msgid "show_type"
 msgstr ""
