+++ src/components/ItaliaTheme/Blocks/HighlightedContent/Skeleton.jsx
