+++ theme/ItaliaTheme/Blocks/_highlitedContent.scss
@@ -1,4 +1,4 @@
-.block.contentInEvidence {
+.block.highlitedContent {
   .empty-selection {
     padding: 0.5em;
     background-color: #eee;
