+++ theme/ItaliaTheme/Print/_blocks.scss
@@ -29,7 +29,7 @@
     margin: 0 auto;
   }
 
-  .block.contentInEvidence {
+  .block.highlitedContent {
     .col-lg-6 {
       flex: 1 1 30%;
       padding: 0;
