+++ theme/ItaliaTheme/Subsites/ItaliaTheme/Blocks/_highlitedContent.scss
@@ -1,4 +1,4 @@
-.block.contentInEvidence {
+.block.highlitedContent {
   .card-title {
     a {
       &:hover,
