+++ theme/ItaliaTheme/Subsites/_mixin.scss
@@ -27,7 +27,7 @@
     @import 'ItaliaTheme/Blocks/cardWithImageAndInEvidence';
     @import 'ItaliaTheme/Blocks/completeBlockLinktemplate';
     @import 'ItaliaTheme/Blocks/ribbonCardTemplate';
-    @import 'ItaliaTheme/Blocks/contentInEvidence';
+    @import 'ItaliaTheme/Blocks/highlitedContent';
     @import 'ItaliaTheme/Blocks/argumentsInEvidence';
     @import 'ItaliaTheme/Blocks/searchSections';
     @import 'ItaliaTheme/Blocks/twitterPosts';
