+++ theme/site.scss
@@ -35,7 +35,7 @@
 @import 'ItaliaTheme/Blocks/calendar';
 @import 'ItaliaTheme/Blocks/alert';
 @import 'ItaliaTheme/Blocks/titolo_argomento';
-@import 'ItaliaTheme/Blocks/contentInEvidence';
+@import 'ItaliaTheme/Blocks/highlitedContent';
 @import 'ItaliaTheme/Blocks/iconBlocks';
 @import 'ItaliaTheme/Blocks/searchSections';
 @import 'ItaliaTheme/Blocks/cardWithImageAndInEvidence';
